+++ Sources/PippinLibrary/Foundation/Set/Set+Factory.swift
@@ -1,0 +1,14 @@
+//
+//  Set+Factory.swift
+//  PippinLibrary
+//
+//  Created by Andrew McKnight on 4/19/20.
+//
+
+import Foundation
+
+public extension Set {
+    static var empty: Set {
+        return Set()
+    }
+}